requirements.txt
fake-useragent
python-dotenv
requests
yt-dlp
clite